@@ -5,18 +5,24 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.3
   flutter_spinkit: ^5.0.0
   http: ^0.13.3
+  json_annotation: ^4.0.1
+  meta: ^1.3.0
+  timeago: ^3.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner:
+  json_serializable: ^4.1.3
+  lints: ^1.0.1
 
 flutter:
   uses-material-design: true
