@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
 import 'package:hackernews/screens/loading_screen.dart';
-import 'package:hackernews/screens/news_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -10,11 +9,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: ThemeData.dark(),
-      routes: {
-        '/load': (context) => LoadingScreen(),
-        '/newsscreen': (context) => NewsScreen(),
-      },
-      initialRoute: '/load',
+      home: LoadingScreen(),
     );
   }
 }
