@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-// import 'package:hackernews/screens/news_screen.dart';
+import 'package:hackernews/screens/news_screen.dart';
 import 'package:hackernews/services/news.dart';
 
 import 'package:flutter_spinkit/flutter_spinkit.dart';
@@ -17,9 +17,14 @@ class _LoadingScreenState extends State<LoadingScreen> {
     getNewsData();
   }
 
-  void getNewsData() async {
-    await NewsModel().getNews();
-    Navigator.pushReplacementNamed(context, '/newsscreen');
+  dynamic getNewsData() async {
+    final newsDataInfo = await NewsModel().getPosts();
+
+    Navigator.push(context, MaterialPageRoute(builder: (context) {
+      return NewsScreen(
+        newsModelScreen: newsDataInfo,
+      );
+    }));
   }
 
   @override
@@ -42,3 +47,6 @@ class _LoadingScreenState extends State<LoadingScreen> {
     );
   }
 }
+
+
+// Navigator.pushReplacementNamed(context, '/newsscreen');
